Utils.py only used by L09.  Move its contents to L09 and flagdefs.py as indicated
